requirements.txt
# Core ML/DL Frameworks
torch>=2.0.0
transformers>=4.30.0
datasets>=2.14.0

# NLP Libraries
spacy>=3.5.0
nltk>=3.8.0
sentencepiece>=0.1.99

# Evaluation Metrics
bert-score>=0.3.13
sacrebleu>=2.3.1
evaluate>=0.4.0
rouge-score>=0.1.2

# Factuality & QA
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
question-generation>=0.1.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
tqdm>=4.65.0

# Configuration & Logging
PyYAML>=6.0
wandb>=0.15.0
tensorboard>=2.13.0

# Utilities
click>=8.1.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Optional: For advanced features
# faiss-gpu>=1.7.4  # Use this instead of faiss-cpu if GPU available
# torch-xla  # For TPU support
